requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
cachetools==4.2.2
celery==5.0.5
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.10.0
Django==3.2.5
django-celery-beat==2.2.0
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
docopt==0.6.2
idna==3.2
kiwisolver==1.3.1
kombu==5.1.0
matplotlib==3.3.3
multidict==5.1.0
numpy==1.21.1
Pillow==8.0.1
prompt-toolkit==3.0.19
psycopg2==2.8.6
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stomp.py==7.0.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
